abstracted flowTypes & propTypes
