+++ lib/Resizable.js
@@ -1,120 +1,14 @@
 // @flow
 import React from 'react';
-import PropTypes from 'prop-types';
-import {DraggableCore, DraggableCoreProps} from 'react-draggable';
+import type {Node as ReactNode} from 'react';
+import {DraggableCore} from 'react-draggable';
+
 import {cloneElement} from './utils';
-import type {Element as ReactElement, Node as ReactNode, ElementRef} from 'react';
+import {resizableProps} from "./propTypes";
+import type {ResizeHandle, Props, ResizableState, DragCallbackData} from './propTypes';
 
-type Axis = 'both' | 'x' | 'y' | 'none';
-type ResizeHandle = 's' | 'w' | 'e' | 'n' | 'sw' | 'nw' | 'se' | 'ne';
-type State = {
-  slackW: number, slackH: number
-};
-type DragCallbackData = {
-  node: HTMLElement,
-  x: number, y: number,
-  deltaX: number, deltaY: number,
-  lastX: number, lastY: number
-};
-export type ResizeCallbackData = {
-  node: HTMLElement,
-  size: {width: number, height: number},
-  handle: ResizeHandle
-};
-export type Props = {
-  children: ReactElement<any>,
-  className?: ?string,
-  width: number,
-  height: number,
-  handle: ReactElement<any> | (resizeHandle: ResizeHandle) => ReactElement<any>,
-  handleSize: [number, number],
-  lockAspectRatio: boolean,
-  axis: Axis,
-  minConstraints: [number, number],
-  maxConstraints: [number, number],
-  onResizeStop?: ?(e: SyntheticEvent<>, data: ResizeCallbackData) => any,
-  onResizeStart?: ?(e: SyntheticEvent<>, data: ResizeCallbackData) => any,
-  onResize?: ?(e: SyntheticEvent<>, data: ResizeCallbackData) => any,
-  draggableOpts?: ?DraggableCoreProps,
-  resizeHandles: ResizeHandle[],
-  transformScale: number,
-};
-
-export default class Resizable extends React.Component<Props, State> {
-  static propTypes = {
-    //
-    // Required Props
-    //
-
-    // Require that one and only one child be present.
-    children: PropTypes.element.isRequired,
-
-    // Initial w/h
-    width: PropTypes.number.isRequired,
-    height: PropTypes.number.isRequired,
-
-    //
-    // Optional props
-    //
-
-    // Custom resize handle
-    handle: PropTypes.oneOfType([
-      PropTypes.node,
-      PropTypes.func
-    ]),
-    // If you change this, be sure to update your css
-    handleSize: PropTypes.arrayOf(PropTypes.number),
-
-    // Defines which resize handles should be rendered (default: 'se')
-    // Allows for any combination of:
-    // 's' - South handle (bottom-center)
-    // 'w' - West handle (left-center)
-    // 'e' - East handle (right-center)
-    // 'n' - North handle (top-center)
-    // 'sw' - Southwest handle (bottom-left)
-    // 'nw' - Northwest handle (top-left)
-    // 'se' - Southeast handle (bottom-right)
-    // 'ne' - Northeast handle (top-center)
-    resizeHandles: PropTypes.arrayOf(PropTypes.oneOf(['s', 'w', 'e', 'n', 'sw', 'nw', 'se', 'ne'])),
-    transformScale: PropTypes.number,
-
-    // If true, will only allow width/height to move in lockstep
-    lockAspectRatio: PropTypes.bool,
-
-    // Restricts resizing to a particular axis (default: 'both')
-    // 'both' - allows resizing by width or height
-    // 'x' - only allows the width to be changed
-    // 'y' - only allows the height to be changed
-    // 'none' - disables resizing altogether
-    axis: PropTypes.oneOf(['both', 'x', 'y', 'none']),
-
-    // Min/max size
-    minConstraints: PropTypes.arrayOf(PropTypes.number),
-    maxConstraints: PropTypes.arrayOf(PropTypes.number),
-
-    // Callbacks
-    onResizeStop: PropTypes.func,
-    onResizeStart: PropTypes.func,
-    onResize: PropTypes.func,
-
-    // These will be passed wholesale to react-draggable's DraggableCore
-    draggableOpts: PropTypes.shape({
-      allowAnyClick: PropTypes.bool,
-      cancel: PropTypes.string,
-      children: PropTypes.node,
-      disabled: PropTypes.bool,
-      enableUserSelectHack: PropTypes.bool,
-      offsetParent: PropTypes.node,
-      grid: PropTypes.arrayOf(PropTypes.number),
-      handle: PropTypes.string,
-      nodeRef: PropTypes.object,
-      onStart: PropTypes.func,
-      onDrag: PropTypes.func,
-      onStop: PropTypes.func,
-      onMouseDown: PropTypes.func,
-      scale: PropTypes.number,
-    }),
-  };
+export default class Resizable extends React.Component<Props, ResizableState> {
+  static propTypes = resizableProps;
 
   static defaultProps =  {
     handleSize: [20, 20],
@@ -126,7 +20,7 @@ export default class Resizable extends React.Component<Props, State> {
     transformScale: 1
   };
 
-  state: State = {
+  state: ResizableState = {
     slackW: 0, slackH: 0,
   };
