+++ lib/ResizableBox.js
@@ -1,68 +1,22 @@
 // @flow
 import React from 'react';
-import PropTypes from 'prop-types';
-import Resizable from './Resizable';
-import type {Props as ResizableProps, ResizeCallbackData} from './Resizable';
 import type {Node as ReactNode} from 'react';
 
-type State = {
-  width: number, height: number,
-  propsWidth: number, propsHeight: number
-};
+import Resizable from './Resizable';
+import {resizableProps} from "./propTypes";
+import type {Props as ResizableProps, ResizeCallbackData, ResizableBoxState} from './propTypes';
 
-export default class ResizableBox extends React.Component<ResizableProps, State> {
-  static propTypes = {
-    children: PropTypes.node.isRequired,
-    height: PropTypes.number.isRequired,
-    width: PropTypes.number.isRequired,
-    handle: PropTypes.oneOfType([
-      PropTypes.node,
-      PropTypes.func,
-    ]),
-    handleSize: PropTypes.arrayOf(PropTypes.number),
-    onResize: PropTypes.func,
-    onResizeStart: PropTypes.func,
-    onResizeStop: PropTypes.func,
-    draggableOpts: PropTypes.shape({
-      allowAnyClick: PropTypes.bool,
-      cancel: PropTypes.string,
-      children: PropTypes.node,
-      disabled: PropTypes.bool,
-      enableUserSelectHack: PropTypes.bool,
-      offsetParent: PropTypes.node,
-      grid: PropTypes.arrayOf(PropTypes.number),
-      handle: PropTypes.string,
-      nodeRef: PropTypes.object,
-      onStart: PropTypes.func,
-      onDrag: PropTypes.func,
-      onStop: PropTypes.func,
-      onMouseDown: PropTypes.func,
-      scale: PropTypes.number,
-    }),
-    minConstraints: PropTypes.arrayOf(PropTypes.number),
-    maxConstraints: PropTypes.arrayOf(PropTypes.number),
-    lockAspectRatio: PropTypes.bool,
-    axis: PropTypes.oneOf(['both', 'x', 'y', 'none']),
-    resizeHandles: PropTypes.arrayOf(PropTypes.oneOf(['s', 'w', 'e', 'n', 'sw', 'nw', 'se', 'ne'])),
-  };
+export default class ResizableBox extends React.Component<ResizableProps, ResizableBoxState> {
+  static propTypes = resizableProps;
 
-  static defaultProps = {
-    handleSize: [20, 20],
-    lockAspectRatio: false,
-    axis: 'both',
-    minConstraints: [20, 20],
-    maxConstraints: [Infinity, Infinity],
-    resizeHandles: ['se'],
-  };
-
-  state: State = {
+  state: ResizableBoxState = {
     width: this.props.width,
     height: this.props.height,
     propsWidth: this.props.width,
     propsHeight: this.props.height,
   };
 
-  static getDerivedStateFromProps(props: ResizableProps, state: State) {
+  static getDerivedStateFromProps(props: ResizableProps, state: ResizableBoxState) {
     // If parent changes height/width, set that in our state.
     if (state.propsWidth !== props.width || state.propsHeight !== props.height) {
       return {
@@ -103,24 +57,26 @@ export default class ResizableBox extends React.Component<ResizableProps, State>
       width,
       height,
       resizeHandles,
+      transformScale,
       ...props
     } = this.props;
 
     return (
       <Resizable
+        axis={axis}
+        draggableOpts={draggableOpts}
         handle={handle}
         handleSize={handleSize}
-        width={this.state.width}
         height={this.state.height}
+        lockAspectRatio={lockAspectRatio}
+        maxConstraints={maxConstraints}
+        minConstraints={minConstraints}
         onResizeStart={onResizeStart}
         onResize={this.onResize}
         onResizeStop={onResizeStop}
-        draggableOpts={draggableOpts}
-        minConstraints={minConstraints}
-        maxConstraints={maxConstraints}
-        lockAspectRatio={lockAspectRatio}
-        axis={axis}
         resizeHandles={resizeHandles}
+        transformScale={transformScale}
+        width={this.state.width}
       >
         <div style={{width: this.state.width + 'px', height: this.state.height + 'px'}} {...props} />
       </Resizable>
